% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_route.R
\name{create_routes}
\alias{create_routes}
\title{Create Routes}
\usage{
create_routes(
  df,
  home_lat,
  home_long,
  home_label,
  station_label,
  dist_from_home
)
}
\arguments{
\item{df}{}

\item{home_lat}{}

\item{home_long}{}

\item{home_label}{}

\item{station_label}{}

\item{dist_from_home}{}
}
\value{
leaflet map showing points connected via routes
}
\description{
Create Routes
}
