% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_fuelwatch}
\alias{download_fuelwatch}
\title{Download the data}
\usage{
download_fuelwatch(
  start_year,
  end_year = Sys.Date(),
  temp_path = here::here("data"),
  base_url =
    "https://warsydprdstafuelwatch.blob.core.windows.net/historical-reports/FuelWatchRetail-"
)
}
\arguments{
\item{start_year}{start year to download data}

\item{end_year}{end year to download data}

\item{temp_path}{path to store data}

\item{base_url}{fuelwatch website for data extraction (end point)}
}
\value{
downloads csv data from Fuelwatch website and writes a dataframe.
  directory
}
\description{
Download fuel watch data from the website
}
\examples{
dontrun{
postcodes <- download_fuelwatch()
}
}
