% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_postcodes}
\alias{download_postcodes}
\title{Download Postcode data}
\usage{
download_postcodes()
}
\value{
downloads an RDS file from github repo and extracts a dataframe
}
\description{
Download Postcode data
}
\examples{
dontrun{
postcodes <- download_postcodes()
}

}
