% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stations.R
\name{find_stations}
\alias{find_stations}
<<<<<<< HEAD
\title{Find Stations}
=======
<<<<<<< HEAD
\title{Find Stations}
=======
\title{Title}
>>>>>>> 17163fa23b1624199c087f28d6ae578f0f324036
>>>>>>> 64cdcf9f7b92b97db671645e7f983c34ff1c43af
\usage{
find_stations(
  home_lat,
  home_long,
  label = "Reference Location",
  dist_from_home,
  radius = dist_from_home * 1000
)
}
\arguments{
<<<<<<< HEAD
\item{label}{the reference location to show on the map}

\item{dist_from_home}{set the maximum distance from reference in kilometers}
}
\description{
The following uses postcode lat and long to estimate distances
from the home location. This is not exact and generally assumes a post code's
central location as the lat/long (i.e. center of each suburb). #Find stations
within reference area The following uses postcode lat and long to estimate
distances from the home location. This is not exact and generally assumes a
post code's central location as the lat/long (i.e. center of each suburb)
=======
\item{home_lat}{}

\item{home_long}{}

<<<<<<< HEAD
\item{label}{the reference location to show on the map}

\item{dist_from_home}{set the maximum distance from reference in kilometers}
=======
\item{label}{}

\item{dist_from_home}{}
>>>>>>> 17163fa23b1624199c087f28d6ae578f0f324036

\item{radius}{}
}
\description{
<<<<<<< HEAD
The following uses postcode lat and long to estimate distances from the home
location. This is not exact and generally assumes a post code's central
location as the lat/long (i.e. center of each suburb)
=======
Title
>>>>>>> 17163fa23b1624199c087f28d6ae578f0f324036
>>>>>>> 64cdcf9f7b92b97db671645e7f983c34ff1c43af
}
