% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_fuel_data}
\alias{get_fuel_data}
\title{Get Fuel Data}
\usage{
get_fuel_data()
}
\value{
a dataframe of fuel price data
}
\description{
Extract the fuel data either from the local project folder or
  download from fuel watch.
}
\details{
This function first checks to see if the RDS file is available and
  opens it. It then checks the age of the data, and if more than two days
  old, downloads the remaining data up to the current month. \cr \cr If there
  is not RDS file, the raw CSV files are checked and opened, and any missing
  are downloaded up to the current month. \cr \cr If there are no CSV files,
  all data is downloaded from the fuelwatch website. Note that this can take
  considerable time and data useage
}
\examples{
dontrun{
fuel <- get_fuel_data()
}

}
