% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos_lat_long.R
\name{get_pos}
\alias{get_pos}
\title{Get Position (Latitude and Longitude)}
\usage{
get_pos(full_address)
}
\arguments{
\item{full_address}{}
}
\value{
a dataframe containing the lat and long of the address provided
}
\description{
Return a Latitude and Longitude for an address. This is a
wrapper for the tmaptools package
}
\examples{
lat_lng <- get_pos("Liberty Island, New York, NY 10004")

}
