% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_postcodes}
\alias{get_postcodes}
\title{Get Postcodes}
\usage{
get_postcodes()
}
\value{
a dataframe of postcode data
}
\description{
Open the pre-saved postcode RDS file or download it from github
if it doesn't exist.

Postcodes are used in geolocation functions of the package.
}
\examples{

dontrun{
postcodes <- get_postcodes
}

}
