% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stations.R
\name{find_stations}
\alias{find_stations}
\title{Find Stations}
\usage{
find_stations(
  home_lat,
  home_long,
  label = "Reference Location",
  dist_from_home,
  radius = dist_from_home * 1000
)
}
\arguments{
\item{label}{the reference location to show on the map}

\item{dist_from_home}{set the maximum distance from reference in kilometers}
}
\description{
The following uses postcode lat and long to estimate distances
from the home location. This is not exact and generally assumes a post code's
central location as the lat/long (i.e. center of each suburb). #Find stations
within reference area The following uses postcode lat and long to estimate
distances from the home location. This is not exact and generally assumes a
post code's central location as the lat/long (i.e. center of each suburb)
}
