% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stations.R
\name{find_stations}
\alias{find_stations}
\title{Title}
\usage{
find_stations(
  home_lat,
  home_long,
  label = "Reference Location",
  dist_from_home,
  radius = dist_from_home * 1000
)
}
\arguments{
\item{home_lat}{}

\item{home_long}{}

\item{label}{}

\item{dist_from_home}{}

\item{radius}{}
}
\description{
Title
}
